+++ benchmarks/benchmarks-utils.mjs
@@ -1,11 +1,14 @@
 import crypto from 'node:crypto'
+import assert from 'node:assert'
 import fs from 'node:fs'
+import Benchmark from 'benchmark'
 import {
   DynamicClusterPool,
   DynamicThreadPool,
   FixedClusterPool,
   FixedThreadPool,
   PoolTypes,
+  WorkerChoiceStrategies,
   WorkerTypes
 } from '../lib/index.mjs'
 import { TaskFunctions } from './benchmarks-types.mjs'
@@ -54,11 +57,11 @@ export const buildPoolifierPool = (
   }
 }
 
-export const runPoolifierTest = async (
+export const runPoolifierPool = async (
   pool,
   { taskExecutions, workerData }
 ) => {
-  return new Promise((resolve, reject) => {
+  return await new Promise((resolve, reject) => {
     let executions = 0
     for (let i = 1; i <= taskExecutions; i++) {
       pool
@@ -66,28 +69,67 @@ export const runPoolifierTest = async (
         .then(() => {
           ++executions
           if (executions === taskExecutions) {
-            return resolve({ ok: 1 })
+            resolve({ ok: 1 })
           }
           return null
         })
         .catch(err => {
           console.error(err)
-          return reject(err)
+          reject(err)
         })
     }
   })
 }
 
-export const getPoolImplementationName = pool => {
-  if (pool instanceof FixedThreadPool) {
-    return 'FixedThreadPool'
-  } else if (pool instanceof DynamicThreadPool) {
-    return 'DynamicThreadPool'
-  } else if (pool instanceof FixedClusterPool) {
-    return 'FixedClusterPool'
-  } else if (pool instanceof DynamicClusterPool) {
-    return 'DynamicClusterPool'
-  }
+export const runPoolifierPoolBenchmark = async (
+  name,
+  pool,
+  { taskExecutions, workerData }
+) => {
+  return await new Promise((resolve, reject) => {
+    try {
+      const suite = new Benchmark.Suite(name)
+      for (const workerChoiceStrategy of Object.values(
+        WorkerChoiceStrategies
+      )) {
+        for (const enableTasksQueue of [false, true]) {
+          suite.add(
+            `${name}|${workerChoiceStrategy}|${
+              enableTasksQueue ? 'with' : 'without'
+            } tasks queue`,
+            async () => {
+              pool.setWorkerChoiceStrategy(workerChoiceStrategy)
+              pool.enableTasksQueue(enableTasksQueue)
+              assert.strictEqual(
+                pool.opts.workerChoiceStrategy,
+                workerChoiceStrategy
+              )
+              assert.strictEqual(pool.opts.enableTasksQueue, enableTasksQueue)
+              await runPoolifierPool(pool, {
+                taskExecutions,
+                workerData
+              })
+            }
+          )
+        }
+      }
+      suite
+        .on('cycle', event => {
+          console.info(event.target.toString())
+        })
+        .on('complete', async function () {
+          console.info(
+            'Fastest is ' +
+              LIST_FORMATTER.format(this.filter('fastest').map('name'))
+          )
+          await pool.destroy()
+          resolve()
+        })
+        .run({ async: true })
+    } catch (error) {
+      reject(error)
+    }
+  })
 }
 
 export const LIST_FORMATTER = new Intl.ListFormat('en-US', {